Add yes/no counts, make creator of group join
The request to see the votes for a restaurant only showed count(yes),
as an integer.

Change so that we receive

```
[
    { "vote": true, "count": xx},
    { "vote" false, "count": xx}
]
```

Also, when a group was created, the founder of the group did
not automatically get added.

Since they arent in the group they can't see the invite code to share.

This made it impossible to join a group.

Add founder to group upon instantiation of a group.